Add Admin Password Reset
This is a function that was left out from the final product that we
handed in for our project. It was supposed to be part of the
adminViewUser but was ultimately left out due to time constraint.

As well as fixed the customerHomePage.php. Apparently the landing was
not working properly and would not display the username of the user
logged in. It was calling for $_SESSION['username'] which is no where to
be found and have no idea how it got there. Either way, it's been fixed
and now works properly as well as a better sql query that is called.