Organize Files
Something I've been wanting to do for the longest time was reogranize
the files. I took the time to do that and change some of the names as
well. I've added /some/ modularity to the program with the use of
'include.php' and 'customerNavBar.php' to help organize important
includes and a navbar for the customer page